wam_ai_demo/requirements.txt
# WAM AI Demo Requirements
snowflake-snowpark-python>=1.18.0
snowflake-ml-python>=1.5.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
